Changes for settings and sleep (#190)
* ClientTimeshift Settings

Test of internal value would force settings change.  This logic will not be ported to Nexus as the support for backend V4 will be removed in Nexus, simplifying all settings.

* Sleep change

Disable potential server calls after sleep

* Handle timer change from core

A timer can be updated by core during a recording which can start a recording but core doesn't fire the corresponding recording update.   This forces the update even when polling is off will watching a recording.

* Recording specific art

Backend V5.2.2 added a new tag that provides more specific art for shows and movie with the same

Co-authored-by: emveepee <nospam>